Added the Railways Announcement Software program
For creating railway announcement software, we will be using a bunch of modules like pyaudio, pydub, and gTTS to process audio and get the announcing status of thousands of trains. By using PyAudio module, we can easily use Python to play and record audio on a variety of platforms. Pydub is a simple and well-designed Python module for audio manipulation and gTTS (which stands for Google Text-to-Speech) is a Python library and CLI tool to interface with Google Translate text-to-speech API.